Add a test to measure SPF efficiency
The test spawns a T number of threads, each mmap-ing A number of areas
each containing a P number of pages and faulting them in. T, A and P
are parameters of the test:

spf_test <T> <A> <P>

The test measures the total amount of time it took for the threads to
mmap and fault in the memory.

Signed-off-by: Suren Baghdasaryan <surenb@google.com>